Finished updating logic for moving between operating conditions. Ready to test.
Updated the logic to move step-wise between operating conditions. Basically it divides the gap between operating conditions into <=10 steps and moves along a linear input trajectory between steps. Once it gets to a step, it pauses to let the system settle, then collects data at that stationary input before moving on to the next step. Once it hits the final target operating condition, it moves on as before. If the flame blows out, it collects one sample for the blow-out state before moving on to the next operating condition.